fix:index dinamico

+++ src/models/Product.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false,
         },
         price: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.DECIMAL(5,2),
             allowNull: false,
         },
         stock: DataTypes.INTEGER,
